# Optimize inode lock usage and direntry lookup in ext2