--- a/kernel/src/fs/ext2/impl_for_vfs/inode.rs
+++ b/kernel/src/fs/ext2/impl_for_vfs/inode.rs
@@ -29,22 +29,7 @@ impl Inode for Ext2Inode {
     }
 
     fn metadata(&self) -> Metadata {
-        Metadata {
-            dev: 0, // TODO: ID of block device
-            ino: self.ino() as _,
-            size: self.file_size() as _,
-            blk_size: self.fs().super_block().block_size(),
-            blocks: self.blocks_count() as _,
-            atime: self.atime(),
-            mtime: self.mtime(),
-            ctime: self.ctime(),
-            type_: self.inode_type(),
-            mode: InodeMode::from(self.file_perm()),
-            nlinks: self.hard_links() as _,
-            uid: Uid::new(self.uid()),
-            gid: Gid::new(self.gid()),
-            rdev: self.device_id(),
-        }
+        self.metadata()
     }
 
     fn atime(&self) -> Duration {