This PR primarily includes:
1. Less locking in some inode micro-ops by: stripping the lock in `InodeImpl`, place fine-grained locks in a new struct `InodeBlockManager`, decoupled from `InodeDesc`.
2. Slightly improve the `DirEntry` lookup, leaving `DirEntry` updating as TODO which requires more work.

The benefit, taking benchmark `sqlite-speedtest1-ext2` as example:
| Case                 | old (latency) | New     |
| -------------------- | ------------- | ------- |
| SELECTS on an IPK    | 6.388s        | 5.190s  |
| SELECTS on a TEXT PK | 6.762s        | 5.560s  |
| ...                  |               |         |
| TOTAL                | 81.449s       | 75.440s |